changes to accordions per Yoshimi
The font size for Standard Group and Regular Quiet needs to be slightly larger.
Add a few more pixels to the space between the arrow icon and the text.

On Standard Group:
Having both the border and arrow icon on the left doesn't work well. If we keep the border, the arrow icon should move to the right.
I prefer the color combo, cgrey / cgrey-05 over cgrey / cprimary-lt.

On Quiet:
It seems unnecessary to have both a border and an arrow icon. Removed the border.